Inline extra object to StringInterpOpt
[Cherry-picked ed63fe8dfa9e7535ece8ee38ba147bec963a8b93]
